+++ ohio-aes/run.sh
@@ -7,6 +7,7 @@ export MQTT_PORT=$(bashio::config 'mqtt_port')
 export MQTT_USER=$(bashio::config 'mqtt_user')
 export MQTT_PASS=$(bashio::config 'mqtt_pass')
 export MQTT_TOPIC=$(bashio::config 'mqtt_topic')
+export MQTT_TOPIC_HOURLY=$(bashio::config 'mqtt_topic_hourly')
 export DAYS_BACK=$(bashio::config 'days_back')
 export RUN_INTERVAL_HOURS=$(bashio::config 'run_interval_hours')